backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "padel_app"
version = "1.0.0"
description = "Padel backend Flask app"
authors = [
  { name="Pedro Pacheco", email="pedropachecocontas@gmail.com" }
]
requires-python = ">=3.10,<4"

dependencies = [
    # --- Flask core ---
    "Flask==2.3.3",
    "Werkzeug==2.3.7",
    "Jinja2==3.1.4",
    "itsdangerous==2.1.2",
    "click==8.1.7",
    "MarkupSafe==2.1.5",
    "Flask-Cors==4.0.0",
    "Flask-JWT-Extended==4.6.0",

    # --- DB ---
    "SQLAlchemy==1.4.52",
    "Flask-SQLAlchemy==2.5.1",
    "psycopg2-binary==2.9.9",

    # --- Migrations ---
    "Flask-Migrate==4.0.7",
    "alembic==1.13.2",

    # --- Sessions / auth / assets / mail ---
    "Flask-Session==0.6.0",
    "Flask-Login==0.6.3",
    "Flask-Assets==2.1.0",
    "Flask-Mail==0.9.1",
    "pyScss==1.4.0",

    # --- Google Cloud ---
    "google-cloud-storage==3.3.1",
    "google-cloud-secret-manager==2.24.0",

    # --- Misc runtime ---
    "unidecode==1.3.8",
    "typing-extensions>=4.6,<5",
    "python-dotenv>=1.0.1",
    "python-dateutil==2.9.0"
]

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
pytest = "*"
black = "^24.0.0"
flake8 = "^7.0.0"
autoflake = "^2.3.1"
isort = "*"
mypy = "*"

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'
exclude = '''
/(
    migrations
  | \.venv
  | build
  | dist
)/
'''

[tool.setuptools]
packages = ["padel_app"]

[tool.setuptools.package-dir]
padel_app = "padel_app"
